refactor get inbox messages
