+++ server/services/MessageServices.js
@@ -5,7 +5,7 @@ import { Sent } from '../model/sent';
 import { Inbox } from '../model/inbox';
 import { Email } from '../model/email';
 import { Draft } from '../model/draft';
-import query from '../db/index';
+// import query from '../db/index';
 
 export default class messageServices {
   static async saveDraft(data) {
@@ -112,38 +112,8 @@ export default class messageServices {
     };
   }
 
-  static async getUsersMessages(userId) {
-    const dbQuery = `SELECT emails.id as id,  emails.subject as subject, emails.message as message, emails.parentmessageid as parentMessageId,
-    emails.status as status, inboxs.receiverid as receiverId, inboxs.senderid as senderId, inboxs.read as read, inboxs.createdat as createdOn
-    FROM inboxs
-    INNER JOIN emails ON inboxs.messageid = emails.id  WHERE inboxs.receiverid = $1;
-     `;
-    try {
-      const { rows } = await query(dbQuery, [userId]);
-      if (!rows[0]) {
-        return {
-          success: true,
-          data: [{
-            message: 'You don\'t have any Inbox yet',
-          }],
-        };
-      }
-      const data = rows;
-      return {
-        success: true,
-        data: [data],
-      };
-    } catch (error) {
-      return {
-        success: false,
-        error: [error],
-      };
-    }
-  }
-
-
   static async getRecievedEmails(userId) {
-    const response = await this.getUsersMessages(userId);
+    const response = await Email.getUsersMessages(userId);
     if (response.success) {
       return {
         status: 200,
